Revamped Settings + Timer + Timer Status Component
Made all the settings more lively with emojis and cards as sub categories. Love how the design is turning out. Added a timer context to store information about the timer status, how long focus & break sessions should be etc. Added those options in the settings.  Fixed the timer. The Timer wasn't connected to the settings page or anything else properly. Now the timer reacts to the amount of time set in the settings for focus & breaks, if its set to count up or down etc. Added an animated container above the timer that mentions what mode the timer is in currently (focus, break, long break).