SettingsModal logic
Created contexts/providers for all settings options not connected to one.

All settings on the frontend are now passed/saved on the server when Save Settings is pressed, allowing for proper sync between sessions & devices.

Created a new context to handle notifications settings effectively. Also created new context for break and stand up reminders.

Added a nicely designed font picker to the settings modal along with a context. The context value affects the font of the main timer. Certain fonts are now loaded at the app startup

Sidenote: The appearance page seems to be take time to load on scroll... This may be related to the addition of the font picker. If so, I will re implement it through a different method.

Registered all the new options to the admin page.