Django tasks.py
Fixed bug where settings would not be saved to the server if the modal was dismissed differently.

Replacing celery with django q as its integrated with django more tightly and doesn't require a separate message broker. Initially tried to make things work with celery, then moved to django q and spent time installing, properly configuring and fixing bugs with it.

Re wrote tasks.py using django q while fixing bugs with the logic.

Added logic to allow users to recconnect to the same session on the backend.

Added a sync version of the end session method in the CurrentSession model due to a lot of async related problems with getting django q to work.