Sidebar Revamp + Friends System Server Side
Planned django model system for implementing friends.

Added a friends request model, along with a bunch of helper methods to the user data model.

added a time studied today to the timer context. Planned out how to relay this information between the server & the client.

Revamped the sidebar yet again with new animations and style that match the rest of the application. It was looking disconnected beforehand.