Floating SideBar
Tried to revamp the profile tile context menu, tried to incorporate another state checker that would allow for more convenient use, but ended up discarding all the changes due to inefficiencies and bugs.

Revamped the sidebar design. The design is now glassy and floating as opposed to the stuck to the side one like a traditional sidebar we had earlier. Did a lot of work on the sidebar, had to spend a lot of time fixing UI bugs caused due to the revamp.

A lot of work still needs to be done to get the colours just right for both dark as well as light mode. Committing the code to save my progress.