Neighborhood Mode Auth
Added hack club neighbourhood mode option to settings.

Added a button to completely hide the sidebar. Iterated over multiple animations to hide it, but reverted to the simple one for now due to loads of bugs.

Made certain functions on the django side universal. Had to debug a lot of async errors in the process. Added a few fields to django models. Added some logic for friends.